requirements.txt
pytest==8.0.0
requests==2.31.0

python-dotenv==1.0.1

pytest-html==4.1.1
allure-pytest==2.13.2

pytest-xdist==3.5.0